analog-input-to-led: Debug analog potentiometer input
Attached a breadboard and the potentiometer to Arduiono Maker UNO.

I attached the analog input A0 to the 'center' of the Potentiometer, the
'bottom' to GND, and the 'top' to 5V.

I needed to know what the range of the analog input is, so added a
serial monitor output to follow the values from the input.

It turned out that the range of the analog input is 0 - 1023. I can now
use this information to adapt the state machine.