analog-input-to-led.ino: Adapt state machine
I decided to divide the input range into three. The lower third will be
turn on the red LED, the middle third the yellow LED, and upper third
the green LED.

The screenshot is taken from the 'Arduino Learning Guide For Beginner'
which is basically the Arduino - Potentimeter setup I used.